MSDK-2242 Cleanup Quick and OpenQuick
Go through and fix most of the warnings that come up when building quick,
excluding stuff associated with cocos2d and deprecated things elsewhere
in the SDK.

Claim Fixed: MSDK-2242
Reviewed by: abaude

+++ source/QRenderTexture.cpp
@@ -45,8 +45,8 @@ QRenderTexture::QRenderTexture(int w, int h, int eFormat, unsigned int uDepthSte
     this->w = size.width;
     this->h = size.height;
 
-    this->x = x;
-    this->y = y;
+    this->x = (float) x;
+    this->y = (float) y;
 
     static bool normaliseProperties = true;
